wave0d.2(#250): move sessionWatcher + import-scanner to daemon (w/ shims)
[REFACTOR-ONLY] No logic changes. File moves + import path updates +
re-export shims. Continues the Wave 0d series (PR #955 / Task #252
moved pure-logic shared/agent/sessionTitles/notify-deciders).

Moves:
- electron/sessionWatcher/{emitDecider,fileSource,inference,index,
  pendingRenameFlusher,projectKey,stateEmitter,titleEmitter}.ts
  -> packages/daemon/src/sessionWatcher/
- electron/sessionWatcher/__tests__/*.test.ts (10 specs)
  -> packages/daemon/src/sessionWatcher/__tests__/*.spec.ts
- electron/import-scanner.ts
  -> packages/daemon/src/importScanner/import-scanner.ts
- tests/import-scanner.test.ts
  -> packages/daemon/src/importScanner/__tests__/import-scanner.spec.ts

Import-path edits per NodeNext (.js extensions on relative imports);
inference.ts switched from `../../src/shared/sessionState` to
`../shared/sessionState.js` because daemon's `rootDir: src` forbids
crossing into the renderer tree. Added a daemon-local copy of the
SessionState union at packages/daemon/src/shared/sessionState.ts (the
renderer keeps its copy at src/shared/sessionState.ts; v0.4 unifies
via Connect-RPC generated types).

Re-export shims at original electron/ paths so out-of-zone consumers
keep compiling unchanged:
- electron/sessionWatcher/index.ts (consumed by electron/main.ts,
  electron/testHooks.ts, electron/ptyHost/{lifecycle,entryFactory}.ts,
  electron/notify/bootstrap/installPipeline.ts)
- electron/sessionWatcher/projectKey.ts (consumed by
  electron/ptyHost/jsonlResolver.ts)

import-scanner has zero non-test consumers, so no shim needed.

Verification:
- pnpm --filter @ccsm/daemon build  GREEN (tsc -p tsconfig.json)
- pnpm --filter @ccsm/daemon exec vitest run src/sessionWatcher
  src/importScanner  GREEN (11 files, 119 tests, all pass)
- npx tsc -p tsconfig.electron.json --noEmit  GREEN
- pnpm lint  GREEN (turbo cache hits + fresh @ccsm/daemon lint)
- bash tools/lint-no-ipc.sh  PASS
