+++ packages/daemon/src/sessionWatcher/__tests__/facade.spec.ts
@@ -19,7 +19,7 @@ import {
   sessionWatcher,
   type UnwatchedEvent,
   type TitleChangedEvent,
-} from '../index';
+} from '../index.js';
 
 let tmpRoot: string | null = null;
