+++ packages/daemon/src/sessionWatcher/__tests__/fileSource.spec.ts
@@ -12,7 +12,7 @@ import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
-import { FileSource, type FileTick } from '../fileSource';
+import { FileSource, type FileTick } from '../fileSource.js';
 
 let tmpRoot: string;
