+++ packages/daemon/src/sessionWatcher/fileSource.ts
